requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
redis[hiredis]==5.2.1
asyncpg==0.30.0
azure-servicebus==7.14.2
azure-identity==1.21.0
azure-storage-blob==12.25.1
prometheus-client==0.21.1
sentry-sdk[fastapi]==2.27.0
python-json-logger==3.3.0
PyJWT==2.10.1
pydantic-settings==2.9.1

